backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pandas==2.2.3
numpy==2.2.5
scikit-learn==1.6.1
python-multipart==0.0.20
google-generativeai==0.8.5
httpx==0.28.1
scipy==1.15.3
google-cloud-firestore==2.20.2
